Remove GameCommonId
It's not used and won't be used. It was thought to be useful to pause a game and resume it later. The common ID would have allowed all users to select the same game. We can do it without since the full gamestate is kept by in the store of all players (host, guests).