Handle disconnection from guest in guest and in host
Sometimes, the WebSocket provided by WebSocketClient.onClose() does not contain the IP address of the destination which prevents the identification of the corresponding player (which is needed to update its state among other things): it's not clear if it's a bug or not. The workaround is to look at the connections in LocalConnectionIdStore and remove the ones from WebSocketClient.connections(): the remaining ones (multiple disconnection could occur at the same time if it's the server that loses WIFI for instance) are the ones that are disconnected.

In addition:
- A new instance of websocketclient is created each time CommClient.start() is called because an instance of WebSocketClient (from the library) cannot be reused (following a disconnection for instance)
- Add button in WaitingRoom for the Guest to leave the game
- Add button to reconnect to host whenever the Guest is disconnected.
- Add connection status for players in WaitingRoom (connected, disconnected)
- Players that are still in the game but are disconnected (they haven't explicitly left the game) are now displayed in the list instead of being filtered out
- When a Guest is disconnected, the connection status of all other players is updated to "disconnected".
- Add processor for connection errors at the Guest