Always dwitch next "Waiting" player if card on table has same value as card played
The "dwitch" attribute from GameState is NOT used by the DwitchEngine at all. It only exists such that the client (e.g. UI) of GameState can react to a player being dwitched.

DwitchEngine determines whether a player is dwitched based on the last card on the table and the card that has just been played (only way to dwitch a player). The only consequence is the the dwitched player skips a turn: this is determined at the same time so the engine doesn't need to store which player is dwitched in the GameState in order to work.
All players dwitch state is reset when any action is performed since it only matters when determining who the next player is.

e.g. 4 players in the playing order Player1, Player2, Player3, Player4.
Player1 (Playing) plays a card that dwitches Player2. Player2.dwitch is set to true and the engine determines that next player is Player3 (who was Waiting). As soon as Player3 does something like "play a card" or "pick a card", the game state must be updated to reflect the action. We leverage this to reset the dwitch state since it no longer matters anymore.