Hard-code connection ID of host
It simplifies the management of the connection IDs since the host is always connected to itself.
This is done because of the following problem: when losing the connection with last client, the connection ID of the host was cleared from the store and never re-added later. Now it's always in the store.