Fix reset of websocket server after a failure
A WebSocketServer (class from lib WebSocket) cannot be re-used: one cannot call "start()" twice on the same instance.

When an error occurs (typically the tuple (IP src, port src) is already in use by another process, start() will fail), the user can click on "Reconnect" in order to retry the start of the comm server. So far, it was re-using the same instance which leades to an exception being thrown. Now, a factory is used to obtain a new instance every time.